audioio/sock_wire: fix UB (signed left-shift of negative) in i16->ring
sock_wire_i16_to_ring() left-shifted a negative int32 (raw == 0x8000 ->
(int16_t)-32768), which is undefined behaviour. CI's ASan+UBSan job runs
-fno-sanitize-recover=undefined and aborts test_sock_wire on it
(sock_wire.h:78: "left shift of negative value -32768"), even though the
result is correct on x86. Shift in uint32_t then convert back to int32_t:
byte-identical on two's-complement, UB-free. test_sock_wire passes under
ASan/UBSan; full unit suite + daemon build unchanged.

+++ audioio/sock_wire.h
@@ -75,7 +75,10 @@ static inline void sock_wire_wr_u64(uint8_t *p, uint64_t v)
 /* i16 wire sample <-> i32 modem-ring sample */
 static inline int32_t sock_wire_i16_to_ring(uint16_t raw)
 {
-    return ((int32_t)(int16_t) raw) << 16;
+    /* Shift in unsigned then convert back: a signed left-shift of a negative
+     * value (raw == 0x8000 -> -32768) is undefined behaviour (UBSan abort);
+     * the unsigned shift is well-defined and byte-identical on 2's-complement. */
+    return (int32_t)((uint32_t)(int16_t) raw << 16);
 }
 
 static inline int16_t sock_wire_ring_to_i16(int32_t s)
